refactor: constants, file names
